requirements-dev.txt
ipdb
pyramid_debugtoolbar
pyramid_debugtoolbar_dogpile
pyramid_ipython
